requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
confusable-homoglyphs==3.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==2.0.5
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.2.0
django-easy-select2==1.3.4
django-environ==0.4.4
django-extensions==1.9.9
django-redis==4.9.0
django-registration==2.4.1
django-registration-redux==2.4
django-rest-swagger==2.2.0
django-smart-selects==1.5.4
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
eventlet==0.17.4
graphviz==0.8.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
Pillow==5.0.0
psycopg2==2.7.4
pydotplus==2.0.2
pyparsing==2.2.0
python-decouple==3.1
pytz==2017.3
redis==2.10.5
requests==2.18.4
simplejson==3.15.0
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
whitenoise==3.3.1
